requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
protobuf==3.12.2
pycparser==2.20
pymongo==3.10.1
six==1.15.0
Werkzeug==1.0.1
